feat(7): fix override test
