+++ scripts/builder/duplose.test.ts
@@ -2,6 +2,6 @@ import { useBuilder } from "./duplose";
 
 describe("useBuilder", () => {
 	it("preflight", () => {
-
+		useBuilder;
 	});
 });
