+++ test/overridesTypes/currentRequestObject/tsconfig.json
@@ -1,7 +1,7 @@
 {	
-  "extends": "../../tsconfig.json",
+  "extends": "../../../tsconfig.json",
   "compilerOptions": {
-    "baseUrl": "../../",
+    "baseUrl": "../../../",
     "paths": {
 		"@utils/*": ["scripts/utils/*"],
 		"@scripts/*": ["scripts/*"],
