+++ test/overridesTypes/testOverrideTypes.sh
@@ -1,0 +1,3 @@
+tsc -p test/overridesTypes/currentRequestObject/tsconfig.json
+tsc -p test/overridesTypes/extractKey/tsconfig.json
+tsc -p test/overridesTypes/httpMethod/tsconfig.json
